Create annotate.py
This code is created to read the lecture content and parse the text inside it. Clean the text data and fetch the spotlight annotation from dbpedia using REST API. Finally save the fetched data as topic_info.csv. This data will help us add topics to lectures and competencies for the students.